pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "process-engineering-optimizer"
version = "0.1.0"
description = "DOE, modeling, inverse design, and RAG artifact storage for process engineering"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
  "numpy>=1.23",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "scipy>=1.10",
  "scikit-learn>=1.3",
  "statsmodels>=0.14",
  "matplotlib>=3.7",
  "joblib>=1.3",
  "typer>=0.12",
  "rich>=13.7",
]

[project.optional-dependencies]
# Heavier optional deps for EHVI-style multi-objective Bayesian optimization (MOBO)
mobo = [
    "ax-platform>=0.3.7",
    "botorch>=0.11.0",
    "torch>=2.1",
]
opt = ["optuna>=3.6"]
xgb = ["xgboost>=2.0"]
rag_dense = ["sentence-transformers>=2.5"]
bo = ["scikit-optimize>=0.10"]
dev = ["pytest>=8.0", "ruff>=0.6"]

[project.scripts]
peo = "process_optimizer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
